Fix for 1. Infinite loading 2. Button looks not working
Issue:
Floating button doesn't seem to do anything.

Root cause:
The issue is because the shift started in one device (in this case mine) and trying to end in another device (Deputy device).

The location permissions on device were asked once during start of the shift and it was assumed that permissions would have already been given while ending the shift(otherwise user wouldn't hit end shift flow). Since the end shift is triggered from another device, permissions are not checked and results in this issue.

Fix:
1. Added permission checks at both start and end shifts.